Biometric Authentication feature done
