+++ app/src/main/java/uz/androbeck/virtualbank/ui/screens/pin_code/events/PinCodeEvent.kt
@@ -1,4 +1,4 @@
-package uz.androbeck.virtualbank.ui.screens.pin_code
+package uz.androbeck.virtualbank.ui.screens.pin_code.events
 
 interface PinCodeEvent {
     data object PinRegistered : PinCodeEvent
