+++ app/src/main/java/uz/androbeck/virtualbank/ui/screens/pin_code/usecases/RegisterPinUseCase.kt
@@ -1,4 +1,4 @@
-package uz.androbeck.virtualbank.ui.screens.pin_code
+package uz.androbeck.virtualbank.ui.screens.pin_code.usecases
 
 import uz.androbeck.virtualbank.preferences.PreferencesProvider
 import javax.inject.Inject
@@ -7,10 +7,6 @@ class RegisterPinUseCase @Inject constructor(
     private val prefsProvider: PreferencesProvider
 ) {
     fun registerPin(pin: String) {
-        if(prefsProvider.pinCodeReserve.isEmpty()){
-            prefsProvider.pinCode = pin
-        } else {
-            prefsProvider.pinCodeReserve = pin
-        }
+        prefsProvider.pinCode = pin
     }
 }
