+++ app/src/main/java/uz/androbeck/virtualbank/ui/screens/pin_code/viewmodels/BiometricPermissionViewModel.kt
@@ -1,4 +1,4 @@
-package uz.androbeck.virtualbank.ui.screens.pin_code
+package uz.androbeck.virtualbank.ui.screens.pin_code.viewmodels
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -9,7 +9,7 @@ import javax.inject.Inject
 class BiometricPermissionViewModel @Inject constructor(
     private val prefsProvider: PreferencesProvider
 ) : ViewModel() {
-    fun setBiometrics(useBiometric: Boolean){
+    fun setBiometrics(useBiometric: Boolean) {
         prefsProvider.useBiometric = useBiometric
     }
 }
