+++ app/src/main/java/uz/androbeck/virtualbank/ui/screens/pin_code/views/ConfirmPinCodeFragment.kt
@@ -1,4 +1,4 @@
-package uz.androbeck.virtualbank.ui.screens.pin_code
+package uz.androbeck.virtualbank.ui.screens.pin_code.views
 
 import android.animation.ValueAnimator
 import android.os.Handler
@@ -16,10 +16,12 @@ import uz.androbeck.virtualbank.databinding.FragmentConfirmPinCodeBinding
 import uz.androbeck.virtualbank.ui.MainViewModel
 import uz.androbeck.virtualbank.ui.base.BaseFragment
 import uz.androbeck.virtualbank.ui.events.NavGraphEvent
-import uz.androbeck.virtualbank.utils.extentions.gone
+import uz.androbeck.virtualbank.ui.screens.pin_code.utils.BiometricUtils
+import uz.androbeck.virtualbank.ui.screens.pin_code.events.ConfirmPinCodeEvent
+import uz.androbeck.virtualbank.ui.screens.pin_code.viewmodels.ConfirmPinCodeViewModel
 import uz.androbeck.virtualbank.utils.extentions.singleClickable
 import uz.androbeck.virtualbank.utils.extentions.vibrate
-import uz.androbeck.virtualbank.utils.extentions.visible
+import uz.androbeck.virtualbank.utils.extentions.visibleIf
 
 @AndroidEntryPoint
 class ConfirmPinCodeFragment : BaseFragment(R.layout.fragment_confirm_pin_code) {
@@ -53,48 +55,81 @@ class ConfirmPinCodeFragment : BaseFragment(R.layout.fragment_confirm_pin_code) 
                 }
             }
 
+            // Back action
             actionBack.singleClickable {
                 vibrate()
+                confirmPinCodeViewModel.clearReservePinCode()
                 findNavController().navigate(R.id.action_confirmPinCodeFragment_to_pinCodeFragment)
             }
 
+            // Confirm actions
             listOf(btnConfirm, actionConfirm).forEach { button ->
                 button.setOnClickListener {
                     vibrate()
-                    //biometrics check and decide which side
+                    confirmPinCodeViewModel.handlePinCodeCompletion()
                 }
             }
 
+            // Remove digit action
             btnRemove.setOnClickListener {
                 vibrate()
                 confirmPinCodeViewModel.removeLastDigit()
             }
-
         }
     }
 
     override fun observe() {
         confirmPinCodeViewModel.apply {
-            pinCodeList.observe(viewLifecycleOwner) {
-                updatePinCode(it)
+            pinCodeList.observe(viewLifecycleOwner) { updatePinCode(it) }
+            confirmPinCodeEvent.observe(viewLifecycleOwner) { handlePinCodeEvent(it) }
+        }
+    }
+
+    private fun handlePinCodeEvent(event: ConfirmPinCodeEvent) {
+        when (event) {
+            ConfirmPinCodeEvent.PinValidated -> {
+                performPinCodeAnimation {
+                    if (BiometricUtils.isBiometricReady(requireContext())) {
+                        findNavController().navigate(R.id.action_confirmPinCodeFragment_to_biometricPermissionFragment)
+                    } else {
+                        confirmPinCodeViewModel.setBiometrics(false)
+                        navigateWithDelay(NavGraphEvent.Main, 600)
+                    }
+                }
+            }
+
+            ConfirmPinCodeEvent.PinValidationFailed -> {
+                performPinCodeAnimation {
+                    errorPinAnim()
+                    confirmPinCodeViewModel.clearPinCode()
+                }
             }
         }
+    }
+
+    private fun performPinCodeAnimation(onAnimationEnd: () -> Unit) {
+        setButtonsEnabled(false)
+        Handler(Looper.getMainLooper()).postDelayed({
+            confirmPinCodeViewModel.clearPinCode()
+        }, 200)
+        Handler(Looper.getMainLooper()).postDelayed({ checkPinAnim() }, 400)
+        Handler(Looper.getMainLooper()).postDelayed({ setButtonsEnabled(true) }, 1000)
+        Handler(Looper.getMainLooper()).postDelayed(onAnimationEnd, 1000)
     }
 
     private fun updatePinCode(pinCode: MutableList<String>) {
         with(binding) {
             val pinViews = listOf(view01, view02, view03, view04)
             pinViews.forEachIndexed { index, view ->
-                view.setBackgroundResource(if (index < pinCode.size) R.drawable.bg_pin_active else R.drawable.bg_pin_inactive)
+                view.setBackgroundResource(
+                    if (index < pinCode.size) R.drawable.bg_pin_active
+                    else R.drawable.bg_pin_inactive
+                )
             }
 
-            if(pinCode.size == 4){
-                actionConfirm.visible()
-                btnConfirm.visible()
-            } else {
-                actionConfirm.gone()
-                btnConfirm.gone()
-            }
+            val isPinComplete = pinCode.size == 4
+            actionConfirm.visibleIf(isPinComplete)
+            btnConfirm.visibleIf(isPinComplete)
         }
     }
 
@@ -143,13 +178,9 @@ class ConfirmPinCodeFragment : BaseFragment(R.layout.fragment_confirm_pin_code) 
     private fun setButtonsEnabled(enabled: Boolean) {
         with(binding) {
             listOf(
-                btn01, btn02, btn03,
-                btn04, btn05, btn06,
-                btn07, btn08, btn09,
-                btn00, btnRemove
-            ).forEach {
-                it.isEnabled = enabled
-            }
+                btn01, btn02, btn03, btn04, btn05, btn06,
+                btn07, btn08, btn09, btn00, btnRemove
+            ).forEach { it.isEnabled = enabled }
         }
     }
 }
