+++ app/src/main/java/uz/androbeck/virtualbank/ui/screens/pin_code/views/PinCodeFragment.kt
@@ -1,0 +1,260 @@
+package uz.androbeck.virtualbank.ui.screens.pin_code.views
+
+import android.animation.ValueAnimator
+import android.os.Handler
+import android.os.Looper
+import android.view.View
+import android.view.animation.AccelerateDecelerateInterpolator
+import android.view.animation.Interpolator
+import androidx.biometric.BiometricPrompt
+import androidx.fragment.app.activityViewModels
+import androidx.fragment.app.viewModels
+import androidx.navigation.fragment.findNavController
+import by.kirich1409.viewbindingdelegate.viewBinding
+import dagger.hilt.android.AndroidEntryPoint
+import uz.androbeck.virtualbank.R
+import uz.androbeck.virtualbank.databinding.FragmentPinCodeBinding
+import uz.androbeck.virtualbank.preferences.PreferencesProvider
+import uz.androbeck.virtualbank.ui.MainViewModel
+import uz.androbeck.virtualbank.ui.base.BaseFragment
+import uz.androbeck.virtualbank.ui.events.NavGraphEvent
+import uz.androbeck.virtualbank.ui.screens.pin_code.events.PinCodeEvent
+import uz.androbeck.virtualbank.ui.screens.pin_code.viewmodels.PinCodeViewModel
+import uz.androbeck.virtualbank.utils.extentions.*
+import java.util.concurrent.Executors
+import javax.inject.Inject
+
+@AndroidEntryPoint
+class PinCodeFragment : BaseFragment(R.layout.fragment_pin_code) {
+    private val binding by viewBinding(FragmentPinCodeBinding::bind)
+    private val pinCodeViewModel: PinCodeViewModel by viewModels()
+    private val sharedVM: MainViewModel by activityViewModels()
+
+    @Inject
+    lateinit var prefsProvider: PreferencesProvider
+
+    override fun setup() {
+        // Additional setup if needed
+    }
+
+    override fun clicks() {
+        setupButtonClicks()
+        setupFingerprintClick()
+    }
+
+    override fun observe() {
+        observePinCodeList()
+        observeFromRegister()
+        observePinCodeEvent()
+        observeErrorAttempts()
+        observeErrorLogout()
+    }
+
+    private fun setupButtonClicks() {
+        val buttonIds = listOf(
+            binding.btn01 to getString(R.string.str_num_1),
+            binding.btn02 to getString(R.string.str_num_2),
+            binding.btn03 to getString(R.string.str_num_3),
+            binding.btn04 to getString(R.string.str_num_4),
+            binding.btn05 to getString(R.string.str_num_5),
+            binding.btn06 to getString(R.string.str_num_6),
+            binding.btn07 to getString(R.string.str_num_7),
+            binding.btn08 to getString(R.string.str_num_8),
+            binding.btn09 to getString(R.string.str_num_9),
+            binding.btn00 to getString(R.string.str_num_0)
+        )
+
+        buttonIds.forEach { (button, digit) ->
+            button.setOnClickListener {
+                vibrate()
+                pinCodeViewModel.addDigit(digit)
+            }
+        }
+
+        binding.btnRemove.setOnClickListener {
+            vibrate()
+            pinCodeViewModel.removeLastDigit()
+        }
+
+        binding.actionExit.setOnClickListener {
+            vibrate()
+            pinCodeViewModel.handlePinCodeExit()
+            navigateWithDelay(NavGraphEvent.Auth, 500)
+        }
+    }
+
+    private fun setupFingerprintClick() {
+        binding.btnFingerprint.singleClickable {
+            vibrate()
+            promptBiometricAuthentication()
+        }
+    }
+
+    private fun observePinCodeList() {
+        pinCodeViewModel.pinCodeList.observe(viewLifecycleOwner) {
+            updatePinCode(it)
+        }
+    }
+
+    private fun observeFromRegister() {
+        pinCodeViewModel.fromRegister.observe(viewLifecycleOwner) {
+            if (it) {
+                if (pinCodeViewModel.checkBiometrics()) {
+                    promptBiometricAuthentication()
+                    binding.btnFingerprint.visible()
+                } else {
+                    binding.btnFingerprint.gone()
+                }
+                binding.actionExit.visible()
+            } else {
+                binding.actionExit.gone()
+                binding.btnFingerprint.gone()
+            }
+        }
+    }
+
+    private fun observePinCodeEvent() {
+        pinCodeViewModel.pinCodeEvent.observe(viewLifecycleOwner) {
+            handlePinCodeEvent(it)
+        }
+    }
+
+    private fun observeErrorAttempts() {
+        pinCodeViewModel.errorAttempts.observe(viewLifecycleOwner) {
+            binding.errorAttempts.text =
+                if (it >= 1) getString(R.string.pin_mismatch_message, 5 - it) else ""
+        }
+    }
+
+    private fun observeErrorLogout() {
+        pinCodeViewModel.errorLogout.observe(viewLifecycleOwner) {
+            if (it) navigateWithDelay(NavGraphEvent.Auth, 0)
+        }
+    }
+
+    private fun handlePinCodeEvent(event: PinCodeEvent) {
+        when (event) {
+            PinCodeEvent.PinRegistered -> {
+                performPinCodeAnimation()
+                Handler(Looper.getMainLooper()).postDelayed({
+                    findNavController().navigate(R.id.action_pinCodeFragment_to_confirmPinCodeFragment)
+                }, 1200)
+            }
+
+            PinCodeEvent.PinValidated -> {
+                performPinCodeAnimation()
+                navigateWithDelay(NavGraphEvent.Main, 1200)
+            }
+
+            PinCodeEvent.PinValidationFailed -> {
+                performPinCodeAnimation()
+                errorPinAnim()
+                pinCodeViewModel.clearPinCode()
+                binding.errorAttempts.visible()
+            }
+        }
+    }
+
+    private fun performPinCodeAnimation() {
+        setButtonsEnabled(false)
+        Handler(Looper.getMainLooper()).postDelayed({
+            pinCodeViewModel.clearPinCode()
+        }, 200)
+        Handler(Looper.getMainLooper()).postDelayed({
+            checkPinAnim()
+        }, 400)
+        Handler(Looper.getMainLooper()).postDelayed({ setButtonsEnabled(true) }, 1000)
+    }
+
+    private fun updatePinCode(pinCode: MutableList<String>) {
+        with(binding) {
+            val pinViews = listOf(view01, view02, view03, view04)
+            pinViews.forEachIndexed { index, view ->
+                view.setBackgroundResource(
+                    if (index < pinCode.size) R.drawable.bg_pin_active
+                    else R.drawable.bg_pin_inactive
+                )
+            }
+
+            if (pinCode.size == 4) {
+                pinCodeViewModel.handlePinCodeCompletion()
+            }
+        }
+    }
+
+    private fun navigateWithDelay(event: NavGraphEvent, delay: Long) {
+        Handler(Looper.getMainLooper()).postDelayed({
+            sharedVM.setNavGraphEvent(event)
+        }, delay)
+    }
+
+    private fun checkPinAnim() {
+        val pinViews = listOf(binding.view01, binding.view02, binding.view03, binding.view04)
+        pinViews.forEachIndexed { index, view ->
+            Handler(Looper.getMainLooper()).postDelayed({
+                view.setBackgroundResource(R.drawable.bg_pin_active)
+            }, index * 100L)
+        }
+    }
+
+    private fun errorPinAnim(
+        view: View = binding.root,
+        offset: Float = 100f,
+        repeatCount: Int = 4,
+        dampingRatio: Float? = null,
+        duration: Long = 1000L,
+        interpolator: Interpolator = AccelerateDecelerateInterpolator()
+    ) {
+        val defaultDampingRatio = dampingRatio ?: (1f / (repeatCount + 1))
+        val animValues = List((repeatCount * 4) + 1) { i ->
+            when (i % 4) {
+                0 -> 0f
+                1 -> -offset * (1 - defaultDampingRatio * (i / 4))
+                2 -> 0f
+                3 -> offset * (1 - defaultDampingRatio * (i / 4))
+                else -> 0f
+            }
+        }.toFloatArray()
+
+        ValueAnimator.ofFloat(*animValues).apply {
+            addUpdateListener { view.translationX = it.animatedValue as Float }
+            this.interpolator = interpolator
+            this.duration = duration
+            start()
+        }
+    }
+
+    private fun promptBiometricAuthentication() {
+        val biometricPrompt = BiometricPrompt(
+            this,
+            Executors.newSingleThreadExecutor(),
+            object : BiometricPrompt.AuthenticationCallback() {
+                override fun onAuthenticationSucceeded(result: BiometricPrompt.AuthenticationResult) {
+                    super.onAuthenticationSucceeded(result)
+                    navigateWithDelay(NavGraphEvent.Main, 0)
+                    pinCodeViewModel.resetErrorAttempts()
+                }
+            })
+
+        val promptInfo = BiometricPrompt.PromptInfo.Builder()
+            .setTitle(getString(R.string.biometric_prompt_title))
+            .setSubtitle(getString(R.string.biometric_prompt_subtitle))
+            .setNegativeButtonText(getString(R.string.biometric_prompt_cancel))
+            .build()
+
+        biometricPrompt.authenticate(promptInfo)
+    }
+
+    private fun setButtonsEnabled(enabled: Boolean) {
+        with(binding) {
+            listOf(
+                btn01, btn02, btn03,
+                btn04, btn05, btn06,
+                btn07, btn08, btn09,
+                btn00, btnRemove, btnFingerprint
+            ).forEach {
+                it.isEnabled = enabled
+            }
+        }
+    }
+}
