+++ app/src/main/res/layout/fragment_confirm_pin_code.xml
@@ -85,22 +85,6 @@
 
     </LinearLayout>
 
-    <!--Error attempts counter-->
-    <TextView
-        android:id="@+id/errorAttempts"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="20dp"
-        android:layout_marginEnd="20dp"
-        android:gravity="center"
-        android:textColor="#B3261E"
-        android:textSize="14sp"
-        android:visibility="gone"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/pinCodePagePINField" />
-
     <!--Num field-->
     <LinearLayout
         android:id="@+id/pinCodePageNumField"
