+++ app/src/main/res/layout/fragment_pin_code.xml
@@ -35,12 +35,12 @@
         android:id="@+id/pinCodePagePINField"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
+        android:layout_marginTop="20dp"
         android:gravity="center"
         android:orientation="horizontal"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
-        android:layout_marginTop="20dp"
         app:layout_constraintTop_toBottomOf="@+id/pinCodePageEnterPin">
 
         <View
@@ -74,33 +74,33 @@
 
     </LinearLayout>
 
-<!--Error attempts counter-->
+    <!--Error attempts counter-->
     <TextView
         android:id="@+id/errorAttempts"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:textSize="14sp"
         android:layout_marginStart="20dp"
-        android:layout_marginEnd="20dp"
         android:layout_marginTop="20dp"
-        android:textColor="#B3261E"
+        android:layout_marginEnd="20dp"
         android:gravity="center"
-        app:layout_constraintTop_toBottomOf="@id/pinCodePagePINField"
-        app:layout_constraintStart_toStartOf="parent"
+        android:textColor="#B3261E"
+        android:textSize="14sp"
+        android:visibility="gone"
         app:layout_constraintEnd_toEndOf="parent"
-        android:visibility="gone"/>
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/pinCodePagePINField" />
 
     <!--Num field-->
     <LinearLayout
         android:id="@+id/pinCodePageNumField"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
+        android:layout_marginBottom="30dp"
         android:gravity="center"
         android:orientation="vertical"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        android:layout_marginBottom="30dp">
+        app:layout_constraintStart_toStartOf="parent">
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -234,12 +234,12 @@
 
             <ImageButton
                 android:id="@+id/btn_remove"
-                android:contentDescription="@string/remove_pin"
                 android:layout_width="50dp"
                 android:layout_height="50dp"
                 android:layout_marginStart="35dp"
                 android:layout_marginEnd="35dp"
                 android:background="@drawable/pin_number_ripple"
+                android:contentDescription="@string/remove_pin"
                 android:scaleType="center"
                 android:src="@drawable/ic_remove_text" />
 
@@ -259,15 +259,16 @@
                 android:layout_height="50dp"
                 android:layout_marginStart="35dp"
                 android:layout_marginEnd="35dp">
+
                 <ImageButton
                     android:id="@+id/btn_fingerprint"
-                    android:contentDescription="@string/confirm_pin"
                     android:layout_width="50dp"
                     android:layout_height="50dp"
                     android:layout_gravity="center"
                     android:layout_marginStart="35dp"
                     android:layout_marginEnd="35dp"
                     android:background="@drawable/pin_number_ripple"
+                    android:contentDescription="@string/confirm_pin"
                     android:src="@drawable/ic_fingerprint"
                     android:visibility="gone" />
             </FrameLayout>
